Reorganize precompiles
