+++ evm/precompiles/__init__.py
@@ -1,0 +1,4 @@
+from .sha256 import precompile_sha256  # noqa: F401
+from .identity import precompile_identity  # noqa: F401
+from .ecrecover import precompile_ecrecover  # noqa: F401
+from .ripemd160 import precompile_ripemd160  # noqa: F401
